move bulk changelog parsing to changelogService
