+++ backend/src/main/java/de/bund/digitalservice/ris/search/service/ChangelogService.java
@@ -6,7 +6,9 @@ import de.bund.digitalservice.ris.search.exception.ObjectStoreServiceException;
 import de.bund.digitalservice.ris.search.importer.changelog.Changelog;
 import de.bund.digitalservice.ris.search.repository.objectstorage.ObjectStorage;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
 import org.apache.logging.log4j.LogManager;
@@ -101,5 +103,45 @@ public class ChangelogService {
       }
     }
     return changelogs;
+  }
+
+  /**
+   * Detects if a changeAll flag is set in any changelog in a list of changelogs
+   *
+   * @param logs changelogs to be checked for a changeAll flag
+   * @return true if any changelog contains a changeAll flag
+   */
+  public static boolean containsChangeAll(List<Changelog> logs) {
+    return !logs.stream().filter(Changelog::isChangeAll).toList().isEmpty();
+  }
+
+  /**
+   * Merges multiple changelogs into a single changelog. If a file is marked as changed and deleted
+   * in the same changelog, it will be marked as deleted. ChangeAll flags are ignored.
+   *
+   * @param changelogs the list of changelogs to merge
+   * @return the merged changelog
+   */
+  public static Changelog mergeChangelogs(List<Changelog> changelogs) {
+    enum Action {
+      CHANGED,
+      DELETED
+    }
+    Map<String, Action> mergedChanges = new HashMap<>();
+    for (Changelog log : changelogs) {
+      log.getChanged().forEach(filename -> mergedChanges.put(filename, Action.CHANGED));
+      log.getDeleted().forEach(filename -> mergedChanges.put(filename, Action.DELETED));
+    }
+    Changelog mergedChangelog = new Changelog();
+    mergedChanges.forEach(
+        (id, type) -> {
+          if (Action.CHANGED.equals(type)) {
+            mergedChangelog.getChanged().add(id);
+          } else {
+            mergedChangelog.getDeleted().add(id);
+          }
+        });
+
+    return mergedChangelog;
   }
 }
