+++ backend/src/test/java/de/bund/digitalservice/ris/search/unit/service/ChangelogServiceTest.java
@@ -12,6 +12,7 @@ import de.bund.digitalservice.ris.search.service.IndexStatusService;
 import de.bund.digitalservice.ris.search.service.IndexSyncJob;
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
 import org.junit.jupiter.api.Assertions;
@@ -67,5 +68,53 @@ class ChangelogServiceTest {
     List<String> changelogs = changelogService.getNewChangelogsPaths(normsBucket, lastSuccess);
     Assertions.assertEquals(changelogs.toArray()[0], changelogFile1);
     Assertions.assertEquals(changelogs.toArray()[1], changelogFile2);
+  }
+
+  @Test
+  void itDetectsChangeAllChangelogs() {
+    Changelog log1 = new Changelog();
+    log1.setChanged(new HashSet<>(List.of("file")));
+    Changelog log2 = new Changelog();
+    log2.setChangeAll(true);
+    Changelog log3 = new Changelog();
+    log3.setDeleted(new HashSet<>(List.of("deleted")));
+
+    Assertions.assertTrue(ChangelogService.containsChangeAll(List.of(log1, log2, log3)));
+  }
+
+  @Test
+  void itDetectsMissingChangeAllInChangelogs() {
+    Changelog log1 = new Changelog();
+    log1.setChanged(new HashSet<>(List.of("file")));
+    Changelog log3 = new Changelog();
+    log3.setDeleted(new HashSet<>(List.of("deleted")));
+
+    Assertions.assertFalse(ChangelogService.containsChangeAll(List.of(log1, log3)));
+  }
+
+  @Test
+  void itMergesAListOfConsecutiveChangelogsIntoASingleOne() {
+    Changelog log1 = new Changelog();
+    log1.setChanged(new HashSet<>(List.of("obsolete", "changed")));
+    log1.setDeleted(new HashSet<>(List.of("obsolete")));
+    Changelog log2 = new Changelog();
+    log2.setChanged(new HashSet<>(List.of("changed2", "obsolete2")));
+    log2.setDeleted(new HashSet<>(List.of("deleted")));
+    Changelog log3 = new Changelog();
+    log3.setChanged(new HashSet<>(List.of("changed3")));
+    log3.setDeleted(new HashSet<>(List.of("obsolete2", "deleted2")));
+
+    var result = ChangelogService.mergeChangelogs(List.of(log1, log2, log3));
+
+    Assertions.assertEquals(3, result.getChanged().size());
+    Assertions.assertTrue(result.getChanged().contains("changed"));
+    Assertions.assertTrue(result.getChanged().contains("changed2"));
+    Assertions.assertTrue(result.getChanged().contains("changed3"));
+
+    Assertions.assertEquals(4, result.getDeleted().size());
+    Assertions.assertTrue(result.getDeleted().contains("deleted"));
+    Assertions.assertTrue(result.getDeleted().contains("deleted2"));
+    Assertions.assertTrue(result.getDeleted().contains("obsolete"));
+    Assertions.assertTrue(result.getDeleted().contains("obsolete2"));
   }
 }
